Reorganize test_window_manager_tools
