+++ tests/miral/display_reconfiguration.cpp
@@ -34,7 +34,7 @@ Height const display_height{480};
 Rectangle const display_area{{display_left,  display_top},
                              {display_width, display_height}};
 
-struct DisplayConfiguration : TestWindowManagerTools
+struct DisplayConfiguration : mt::TestWindowManagerTools
 {
     Size const initial_window_size{600, 400};
