+++ tests/miral/drag_active_window.cpp
@@ -34,7 +34,7 @@ Height const display_height{480};
 Rectangle const display_area{{display_left,  display_top},
                              {display_width, display_height}};
 
-struct DragActiveWindow : TestWindowManagerTools, WithParamInterface<MirWindowType>
+struct DragActiveWindow : mt::TestWindowManagerTools, WithParamInterface<MirWindowType>
 {
     Size const initial_parent_size{600, 400};
