+++ tests/miral/test_window_manager_tools.h
@@ -22,138 +22,18 @@
 #include "basic_window_manager.h"
 
 #include <miral/canonical_window_manager.h>
-#include <miral/output.h>
 
+#include <mir/shell/surface_specification.h>
 #include <mir/scene/surface_creation_parameters.h>
-#include <mir/shell/display_layout.h>
-#include <mir/shell/focus_controller.h>
-#include <mir/shell/persistent_surface_store.h>
-#include "mir/graphics/display_configuration_observer.h"
-
-#include <mir/test/doubles/stub_session.h>
-#include <mir/test/doubles/stub_surface.h>
-#include "mir/test/doubles/mock_display_configuration.h"
-#include <mir/test/fake_shared.h>
+#include <mir/scene/surface.h>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 
-#include <atomic>
-
-struct StubFocusController : mir::shell::FocusController
+namespace mir
 {
-    void focus_next_session() override {}
-    void focus_prev_session() override {}
-
-    auto focused_session() const -> std::shared_ptr<mir::scene::Session> override { return {}; }
-
-    void set_focus_to(
-        std::shared_ptr<mir::scene::Session> const& /*focus_session*/,
-        std::shared_ptr<mir::scene::Surface> const& /*focus_surface*/) override {}
-
-    auto focused_surface() const -> std::shared_ptr<mir::scene::Surface> override { return {}; }
-
-    void raise(mir::shell::SurfaceSet const& /*windows*/) override {}
-
-    virtual auto surface_at(mir::geometry::Point /*cursor*/) const -> std::shared_ptr<mir::scene::Surface> override
-        { return {}; }
-
-    void set_drag_and_drop_handle(std::vector<uint8_t> const& /*handle*/) override {}
-
-    void clear_drag_and_drop_handle() override {}
-};
-
-struct StubDisplayLayout : mir::shell::DisplayLayout
+namespace test
 {
-    void clip_to_output(mir::geometry::Rectangle& /*rect*/) override {}
-
-    void size_to_output(mir::geometry::Rectangle& /*rect*/) override {}
-
-    bool place_in_output(mir::graphics::DisplayConfigurationOutputId /*id*/, mir::geometry::Rectangle& /*rect*/) override
-        { return false; }
-};
-
-struct StubPersistentSurfaceStore : mir::shell::PersistentSurfaceStore
-{
-    Id id_for_surface(std::shared_ptr<mir::scene::Surface> const& /*surface*/) override { return {}; }
-
-    auto surface_for_id(Id const& /*id*/) const -> std::shared_ptr<mir::scene::Surface> override { return {}; }
-};
-
-struct StubSurface : mir::test::doubles::StubSurface
-{
-    StubSurface(
-        std::string name,
-        MirWindowType type,
-        mir::geometry::Point top_left,
-        mir::geometry::Size size,
-        MirDepthLayer depth_layer)
-        : name_{name}, type_{type}, top_left_{top_left}, size_{size}, depth_layer_{depth_layer}
-    {
-    }
-
-    std::string name() const override { return name_; };
-    MirWindowType type() const override { return type_; }
-
-    mir::geometry::Point top_left() const override { return top_left_; }
-    void move_to(mir::geometry::Point const& top_left) override { top_left_ = top_left; }
-
-    mir::geometry::Size size() const override { return  size_; }
-    void resize(mir::geometry::Size const& size) override { size_ = size; }
-
-    auto state() const -> MirWindowState override { return state_; }
-    auto configure(MirWindowAttrib attrib, int value) -> int override {
-        switch (attrib)
-        {
-        case mir_window_attrib_state:
-            state_ = MirWindowState(value);
-            return state_;
-        default:
-            return value;
-        }
-    }
-
-    bool visible() const override { return  state() != mir_window_state_hidden; }
-
-    auto depth_layer() const -> MirDepthLayer override { return depth_layer_; }
-    void set_depth_layer(MirDepthLayer depth_layer) override { depth_layer_ = depth_layer; }
-
-    std::string name_;
-    MirWindowType type_;
-    mir::geometry::Point top_left_;
-    mir::geometry::Size size_;
-    MirWindowState state_ = mir_window_state_restored;
-    MirDepthLayer depth_layer_;
-};
-
-struct StubStubSession : mir::test::doubles::StubSession
-{
-    mir::frontend::SurfaceId create_surface(
-        mir::scene::SurfaceCreationParameters const& params,
-        std::shared_ptr<mir::frontend::EventSink> const& /*sink*/) override
-    {
-        auto id = mir::frontend::SurfaceId{next_surface_id.fetch_add(1)};
-        auto surface = std::make_shared<StubSurface>(
-            params.name,
-            params.type.value(),
-            params.top_left,
-            params.size,
-            params.depth_layer.is_set() ?
-                params.depth_layer.value()
-                : mir_depth_layer_application);
-        surfaces[id] = surface;
-        return id;
-    }
-
-    std::shared_ptr<mir::scene::Surface> surface(mir::frontend::SurfaceId surface) const override
-    {
-        return surfaces.at(surface);
-    }
-
-private:
-    std::atomic<int> next_surface_id;
-    std::map<mir::frontend::SurfaceId, std::shared_ptr<mir::scene::Surface>> surfaces;
-};
 
 struct MockWindowManagerPolicy : miral::CanonicalWindowManagerPolicy
 {
@@ -180,101 +60,29 @@ struct MockWindowManagerPolicy : miral::CanonicalWindowManagerPolicy
     }
 };
 
-struct TestDisplayConfigurationObserver : mir::ObserverRegistrar<mir::graphics::DisplayConfigurationObserver>
+class TestWindowManagerTools : public testing::Test
 {
-    void register_interest(std::weak_ptr<mir::graphics::DisplayConfigurationObserver> const& o) override
-    {
-        ASSERT_THAT(observer.lock(), testing::IsNull())
-            << "TestDisplayConfigurationObserver does not support multiple observers";
-        observer = o;
-    }
+private:
+    struct Self;
+    std::unique_ptr<Self> self;
 
-    void register_interest(std::weak_ptr<mir::graphics::DisplayConfigurationObserver> const& o, mir::Executor&) override
-    {
-        register_interest(o);
-    }
+public:
+    TestWindowManagerTools();
+    ~TestWindowManagerTools();
 
-    void unregister_interest(mir::graphics::DisplayConfigurationObserver const& o) override
-    {
-        ASSERT_THAT(observer.lock().get(), testing::Eq(&o));
-        observer = std::weak_ptr<mir::graphics::DisplayConfigurationObserver>(); // set to null
-    }
-
-    std::weak_ptr<mir::graphics::DisplayConfigurationObserver> observer;
-};
-
-struct TestWindowManagerTools : testing::Test
-{
-    StubFocusController focus_controller;
-    StubDisplayLayout display_layout;
-    StubPersistentSurfaceStore persistent_surface_store;
-    TestDisplayConfigurationObserver display_configuration_observer;
-    std::shared_ptr<StubStubSession> session{std::make_shared<StubStubSession>()};
-
-    MockWindowManagerPolicy* window_manager_policy{nullptr};
-    miral::WindowManagerTools window_manager_tools{nullptr};
-
-    miral::BasicWindowManager basic_window_manager{
-        &focus_controller,
-        mir::test::fake_shared(display_layout),
-        mir::test::fake_shared(persistent_surface_store),
-        display_configuration_observer,
-        [this](miral::WindowManagerTools const& tools) -> std::unique_ptr<miral::WindowManagementPolicy>
-            {
-                auto policy = std::make_unique<testing::NiceMock<MockWindowManagerPolicy>>(tools);
-                window_manager_policy = policy.get();
-                window_manager_tools = tools;
-                return policy;
-            }
-    };
+    std::shared_ptr<mir::scene::Session> session;
+    MockWindowManagerPolicy* window_manager_policy;
+    miral::WindowManagerTools window_manager_tools;
+    miral::BasicWindowManager basic_window_manager;
 
     static auto create_surface(
         std::shared_ptr<mir::scene::Session> const& session,
-        mir::scene::SurfaceCreationParameters const& params) -> mir::frontend::SurfaceId
-    {
-        // This type is Mir-internal, I hope we don't need to create it here
-        std::shared_ptr<mir::frontend::EventSink> const sink;
-        return session->create_surface(params, sink);
-    }
+        mir::scene::SurfaceCreationParameters const& params) -> mir::frontend::SurfaceId;
 
-    void set_outputs(std::vector<miral::Rectangle> outputs)
-    {
-        auto const display_config = std::make_shared<const mir::test::doubles::MockDisplayConfiguration>();
-        EXPECT_CALL(*display_config, for_each_output(testing::_))
-            .WillOnce(testing::Invoke([outputs](std::function<void(mir::graphics::DisplayConfigurationOutput const&)> func){
-                for (auto i = 0u; i < outputs.size(); i++)
-                {
-                    auto const& rect = outputs[i];
-                    mir::graphics::DisplayConfigurationOutput display_config_output{
-                        mir::graphics::DisplayConfigurationOutputId{(int)i}, // id
-                        mir::graphics::DisplayConfigurationCardId{1}, // card_id
-                        mir::graphics::DisplayConfigurationOutputType::unknown, // type
-                        {mir_pixel_format_abgr_8888}, // pixel_formats
-                        {{rect.size, 60}}, // modes
-                        0, // prefered_mode_index
-                        {rect.size}, // physical_size_mm
-                        true, // connected
-                        true, // used
-                        rect.top_left, // top_left
-                        0, // current_mode_index
-                        mir_pixel_format_abgr_8888, // current_format
-                        mir_power_mode_on, // power_mode
-                        mir_orientation_normal, // orientation
-                        1.0, // scale
-                        mir_form_factor_unknown, // form_factor
-                        mir_subpixel_arrangement_unknown, // subpixel_arrangement
-                        {}, // gamma
-                        mir_output_gamma_unsupported, // gamma_supported
-                        {}, // edid
-                        {}, // custom_logical_size
-                    };
-                    func(display_config_output);
-                }
-            }));
-        auto config_observer = display_configuration_observer.observer.lock();
-        ASSERT_THAT(config_observer, testing::NotNull());
-        config_observer->configuration_applied(display_config);
-    }
+    void set_outputs(std::vector<miral::Rectangle> outputs);
 };
+
+}
+}
 
 #endif //MIRAL_TEST_WINDOW_MANAGER_TOOLS_H
